+++ src/main/java/com/example/restservice/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.example.restservice.controller;
 
+import com.example.restservice.BadRequestException;
 import com.example.restservice.model.User;
 import com.example.restservice.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/users")
 @RequiredArgsConstructor
+@CustomExceptionHandler
 @Tag(name = "Контроллер пользователей", description
           = "Позволяет получать, создавать и удалять пользователей")
 public class UserController {
@@ -29,7 +31,24 @@ public class UserController {
             description = "Выводит всех пользователей и информацию о них"
     )
     public List<User> getAllUsers() {
+        List<User> users = userService.getAllUsers();
+        if (users.isEmpty()) {
+            throw new BadRequestException("Список пользователей пуст");
+        }
         return userService.getAllUsers();
+    }
+
+    @GetMapping("/{id}")
+    @Operation(
+            summary = "Выводит пользователя",
+            description = "Выводит пользователя по его уникальному идентификатору"
+    )
+    public User getUserById(@PathVariable long id) {
+        User user = userService.getUserById(id);
+        if (user == null) {
+            throw new BadRequestException("Пользователь с id " + id + " не найден");
+        }
+        return user;
     }
 
     @PostMapping
@@ -37,6 +56,9 @@ public class UserController {
             summary = "Создает пользователей"
     )
     public List<User> createUsers(@RequestBody List<User> users) {
+        if (users == null || users.isEmpty()) {
+            throw new BadRequestException("Список пользователей не может быть пустым");
+        }
         return userService.createUsers(users);
     }
 
@@ -46,6 +68,10 @@ public class UserController {
             description = "Удаляет пользователя по его id"
     )
     public void deleteUser(@PathVariable Long id) {
+        User user = userService.getUserById(id);
+        if (user == null) {
+            throw new BadRequestException("Пользователь с id " + id + " не найден");
+        }
         userService.deleteUser(id);
     }
 }
